+++ Terraria/Terraria/WorldFile.cs
@@ -1436,7 +1436,7 @@ namespace Terraria
 							}
 							else
 							{
-								tile.slope(b4 - 1);
+								tile.slope((byte)(b4 - 1));
 							}
 						}
 					}
